add asm_direct_pass
trying to write a working version that directly translates to asm so I
can get used to writing asm as well as evaluate the requirements for an
abstraction.

Current problems: Const nodes belong to the start block of the graph, not
to the block they've been created in (which is also the one yComp shows
them in). Also, how to generate jmp instructions?